+++ Configuration/Icons.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+use TYPO3\CMS\Core\Imaging\IconProvider\SvgIconProvider;
+
+return [
+    'tx-googlereviews-reviewlist' => [
+        'provider' => SvgIconProvider::class,
+        'source' => 'EXT:google_reviews/Resources/Public/Icons/Extension.svg',
+    ],
+];